refactor: extract document vector store queries to shared module
Move getDocumentVectorStores function from settings/data.ts to lib/vector-stores/document/queries.ts to enable reuse across the application.
